Extract catalog printing surfaces
